requirements.txt
openai>=1.0.0
gradio>=4.0.0
mcp>=0.1.0
openai-agents>=0.1.0
python-dotenv>=1.0.0

